+++ packages/amplify-codegen/src/commands/models.js
@@ -6,9 +6,7 @@ const { FeatureFlags, pathManager } = require('@aws-amplify/amplify-cli-core');
 const gqlCodeGen = require('@graphql-codegen/core');
 const appSyncDataStoreCodeGen = require('@aws-amplify/appsync-modelgen-plugin');
 const { version: packageVersion } = require('../../package.json');
-const { validateDartSDK } = require('../utils/validateDartSDK');
-const { validateAmplifyFlutterCapableZeroThreeFeatures } = require('../utils/validateAmplifyFlutterCapableZeroThreeFeatures');
-const { validateAmplifyFlutterCoreLibraryDependency } = require('../utils/validateAmplifyFlutterCoreLibraryDependency');
+const { validateAmplifyFlutterMinSupportedVersion } = require('../utils/validateAmplifyFlutterMinSupportedVersion');
 
 const platformToLanguageMap = {
   android: 'java',
@@ -98,6 +96,12 @@ async function generateModels(context, generateOptions = null) {
   const schema = parse(schemaContent);
   const projectConfig = context.amplify.getProjectConfig();
 
+  if (!isIntrospection && projectConfig.frontend === 'flutter' && !validateAmplifyFlutterMinSupportedVersion(projectRoot)) {
+    context.print.error(`🚫 Models are not generated!
+Amplify Flutter versions prior to 0.6.0 are no longer supported by codegen. Please upgrade to use codegen.`);
+    return;
+  }
+
   const generateIndexRules = readFeatureFlag('codegen.generateIndexRules');
   const emitAuthProvider = readFeatureFlag('codegen.emitAuthProvider');
   const usePipelinedTransformer = readFeatureFlag('graphQLTransformer.useExperimentalPipelinedTransformer');
@@ -106,29 +110,6 @@ async function generateModels(context, generateOptions = null) {
   const generateModelsForLazyLoadAndCustomSelectionSet = readFeatureFlag('codegen.generateModelsForLazyLoadAndCustomSelectionSet');
 
   let isTimestampFieldsAdded = readFeatureFlag('codegen.addTimestampFields');
-  let enableDartNullSafety = readFeatureFlag('codegen.enableDartNullSafety');
-  let enableDartZeroThreeFeatures = false;
-  let dartUpdateAmplifyCoreDependency = false;
-
-  if (projectConfig.frontend === 'flutter') {
-    const isMinimumDartVersionSatisfied = validateDartSDK(context, projectRoot);
-    context.print.warning(`Detected feature flag: “enableDartNullSafety : ${enableDartNullSafety}”`);
-    if (isMinimumDartVersionSatisfied && enableDartNullSafety) {
-      context.print.warning(
-        'Generating Dart Models with null safety. To opt out of null safe models, turn off the “enableDartNullSafety” feature flag. Learn more: https://docs.amplify.aws/lib/project-setup/null-safety/q/platform/flutter',
-      );
-    } else {
-      enableDartNullSafety = false;
-      context.print.warning(
-        'Generating Dart Models without null safety. To generate null safe data models, turn on the “enableDartNullSafety” feature flag and set your Dart SDK version to “>= 2.12.0”. Learn more: https://docs.amplify.aws/lib/project-setup/null-safety/q/platform/flutter',
-      );
-    }
-    // override isTimestampFieldsAdded to true when using amplify-flutter > 0.3.0 || > 0.3.0-rc.2
-    isTimestampFieldsAdded = validateAmplifyFlutterCapableZeroThreeFeatures(projectRoot);
-    enableDartZeroThreeFeatures = validateAmplifyFlutterCapableZeroThreeFeatures(projectRoot);
-    // This feature is supported only for users using amplify-flutter > 0.4.0 || > 0.4.0-rc.1
-    dartUpdateAmplifyCoreDependency = validateAmplifyFlutterCoreLibraryDependency(projectRoot);
-  }
 
   const handleListNullabilityTransparently = readFeatureFlag('codegen.handleListNullabilityTransparently');
   const appsyncLocalConfig = await appSyncDataStoreCodeGen.preset.buildGeneratesSection({
@@ -140,12 +121,9 @@ async function generateModels(context, generateOptions = null) {
       isTimestampFieldsAdded,
       emitAuthProvider,
       generateIndexRules,
-      enableDartNullSafety,
       handleListNullabilityTransparently,
       usePipelinedTransformer,
-      enableDartZeroThreeFeatures,
       transformerVersion,
-      dartUpdateAmplifyCoreDependency,
       respectPrimaryKeyAttributesOnConnectionField,
       generateModelsForLazyLoadAndCustomSelectionSet,
       codegenVersion: packageVersion,