+++ packages/amplify-codegen-e2e-core/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.7](https://github.com/aws-amplify/amplify-codegen/compare/@aws-amplify/amplify-codegen-e2e-core@1.4.6...@aws-amplify/amplify-codegen-e2e-core@1.4.7) (2023-05-04)
+
+**Note:** Version bump only for package @aws-amplify/amplify-codegen-e2e-core
+
 ## [1.4.6](https://github.com/aws-amplify/amplify-codegen/compare/@aws-amplify/amplify-codegen-e2e-core@1.4.5...@aws-amplify/amplify-codegen-e2e-core@1.4.6) (2023-04-12)
 
 **Note:** Version bump only for package @aws-amplify/amplify-codegen-e2e-core